:truck: move CertifiedScore to certification context
