+++ api/src/certification/evaluation/domain/services/scoring/scoring-v2.js
@@ -17,12 +17,12 @@ import {
   AnswerCollectionForScoring,
   CertificationAssessmentScore,
   CertificationContract,
-  CertifiedScore,
   CompetenceMark,
   ReproducibilityRate,
 } from '../../../../../shared/domain/models/index.js';
 import { AssessmentResultFactory } from '../../../../scoring/domain/models/factories/AssessmentResultFactory.js';
 import { AlgorithmEngineVersion } from '../../../../shared/domain/models/AlgorithmEngineVersion.js';
+import { CertifiedScore } from '../../../domain/models/CertifiedScore.js';
 import { CertifiedLevel } from '../../models/CertifiedLevel.js';
 
 /**
