+++ api/tests/certification/evaluation/unit/domain/models/CertifiedScore_test.js
@@ -1,7 +1,7 @@
-import { CertifiedLevel } from '../../../../src/certification/evaluation/domain/models/CertifiedLevel.js';
-import { PIX_COUNT_BY_LEVEL } from '../../../../src/shared/domain/constants.js';
-import { CertifiedScore } from '../../../../src/shared/domain/models/CertifiedScore.js';
-import { expect } from '../../../test-helper.js';
+import { CertifiedLevel } from '../../../../../../src/certification/evaluation/domain/models/CertifiedLevel.js';
+import { CertifiedScore } from '../../../../../../src/certification/evaluation/domain/models/CertifiedScore.js';
+import { PIX_COUNT_BY_LEVEL } from '../../../../../../src/shared/domain/constants.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | CertifiedScore', function () {
   it('is equal to the estimated score if the estimated level is certified', function () {
